[GM-142] Depreciating Gamut modal component
