+++ packages/gamut/src/Dialog/__tests__/Dialog-test.tsx
@@ -24,7 +24,7 @@ const defaultProps = {
 
 const renderView = setupRtl(Dialog, defaultProps);
 
-describe('Modal', () => {
+describe('DeprecatedModal', () => {
   beforeEach(() => {
     jest.resetAllMocks();
   });
