+++ packages/gamut/src/Interstitial/index.tsx
@@ -8,7 +8,7 @@ export type InterstitialProps = {
   className?: string;
   /** An optional image, emoji, or component that will render above the title. */
   decoration?: React.ReactNode;
-  /** If enabled, focus will be pulled onto the title of the component upon render. Should be enabled if Interstitial is not the child of a component with focus managment, such as Modal. */
+  /** If enabled, focus will be pulled onto the title of the component upon render. Should be enabled if Interstitial is not the child of a component with focus managment, such as DeprecatedModal. */
   focus?: boolean;
   /** h1 title for the interstitial */
   title: string;
