+++ packages/gamut/src/index.tsx
@@ -34,7 +34,7 @@ export * from './Markdown';
 export * from './Menus/MenuItem';
 export * from './Menus/SideMenu';
 export * from './Dialog';
-export * from './Modal';
+export * from './DeprecatedModal';
 export * from './NotificationList';
 export * from './NotificationList/NotificationItem';
 export * from './Overlay';
