+++ packages/styleguide/stories/Molecules/Dialog.stories.mdx
@@ -17,7 +17,7 @@ import { LinkTo, PropsTable } from '~styleguide/blocks';
   args={{
     inline: true,
     isOpen: true,
-    title: 'Depeche Modal',
+    title: 'Depeche DeprecatedModal',
     children: 'All I ever wanted, all I ever needed is here in my',
     confirmCta: { children: 'Arms!' },
     cancelCta: { children: 'Heart?' },
