+++ packages/styleguide/stories/Molecules/Modal.stories.mdx
@@ -1,4 +1,8 @@
-import { ButtonDeprecated, Modal, Overlay } from '@codecademy/gamut/src';
+import {
+  ButtonDeprecated,
+  DeprecatedModal,
+  Overlay,
+} from '@codecademy/gamut/src';
 import title from '@codecademy/macros/lib/title.macro';
 import { PropsTable } from '@codecademy/storybook-addon-variance';
 import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
@@ -7,10 +11,10 @@ import { useState } from 'react';
 
 <Meta
   title={title}
-  component={Modal}
+  component={DeprecatedModal}
   subcomponents={{ Overlay }}
   parameters={{
-    subtitle: `Simple Modal portal to escape current DOM Context.`,
+    subtitle: `Simple DeprecatedModal portal to escape current DOM Context.`,
     status: 'deprecated',
     source: 'gamut',
     figmaId: 'yt7NdZTETDRPUn0CtRu3v0',
@@ -18,7 +22,7 @@ import { useState } from 'react';
 />
 
 <Canvas>
-  <Story name="Modal">
+  <Story name="DeprecatedModal">
     {(args) => {
       const [isOpen, setIsOpen] = useState(false);
       return (
@@ -26,9 +30,12 @@ import { useState } from 'react';
           <ButtonDeprecated onClick={() => setIsOpen(true)}>
             Open
           </ButtonDeprecated>
-          <Modal isOpen={isOpen} onRequestClose={() => setIsOpen(false)}>
-            Close the Modal!
-          </Modal>
+          <DeprecatedModal
+            isOpen={isOpen}
+            onRequestClose={() => setIsOpen(false)}
+          >
+            Close the DeprecatedModal!
+          </DeprecatedModal>
         </>
       );
     }}
@@ -37,15 +44,15 @@ import { useState } from 'react';
 
 > Prefer <LinkTo kind="Molecules/Dialog" story="Dialog">Dialog</LinkTo> instead.
 
-<PropsTable components={{ Modal, Overlay }} />
+<PropsTable components={{ DeprecatedModal, Overlay }} />
 
 ### Creating a custom close button
 
-By default, the Modal will render a close button.
+By default, the DeprecatedModal will render a close button.
 
 You can disable the default close button with the `hideDefaultCloseButton` prop.
 
-You can then use the `isOpen` prop to create other controls that toggle the Modal on and off.
+You can then use the `isOpen` prop to create other controls that toggle the DeprecatedModal on and off.
 
 <Canvas>
   <Story name="Without Close Default">
@@ -56,16 +63,16 @@ You can then use the `isOpen` prop to create other controls that toggle the Moda
           <ButtonDeprecated onClick={() => setIsOpen(true)}>
             Open
           </ButtonDeprecated>
-          <Modal
+          <DeprecatedModal
             isOpen={isOpen}
             hideDefaultCloseButton
             onRequestClose={() => setIsOpen(false)}
           >
-            <p>Close the Modal...</p>
+            <p>Close the DeprecatedModal...</p>
             <ButtonDeprecated block onClick={() => setIsOpen(false)}>
               with this button instead
             </ButtonDeprecated>
-          </Modal>
+          </DeprecatedModal>
         </>
       );
     }}
@@ -74,12 +81,12 @@ You can then use the `isOpen` prop to create other controls that toggle the Moda
 
 ### Alternative closing methods
 
-Other than creating buttons that can close the Modal, there are other ways a Modal can be closed by the user.
+Other than creating buttons that can close the DeprecatedModal, there are other ways a DeprecatedModal can be closed by the user.
 
-1. Clicking the content outside of the Modal
+1. Clicking the content outside of the DeprecatedModal
 2. Clicking the escape key on their keyboard
 
-All of these methods default to true for accessibility reasons, and rely on setting the required `onRequestClose` prop and making sure it toggles the `isOpen` prop on the Modal.
+All of these methods default to true for accessibility reasons, and rely on setting the required `onRequestClose` prop and making sure it toggles the `isOpen` prop on the DeprecatedModal.
 
 <Canvas>
   <Story name="Clicking Outside">
@@ -90,16 +97,16 @@ All of these methods default to true for accessibility reasons, and rely on sett
           <ButtonDeprecated onClick={() => setIsOpen(true)}>
             Open
           </ButtonDeprecated>
-          <Modal
+          <DeprecatedModal
             isOpen={isOpen}
             hideDefaultCloseButton
             onRequestClose={() => setIsOpen(false)}
           >
             <p>
-              Close this Modal by clicking the page outside of the Modal
-              content, or by clicking the escape key.
+              Close this DeprecatedModal by clicking the page outside of the
+              DeprecatedModal content, or by clicking the escape key.
             </p>
-          </Modal>
+          </DeprecatedModal>
         </>
       );
     }}
